requirements.txt
streamlit
tensorflow
numpy
pandas
matplotlib
opencv-python
Pillow
scikit-learn